[PartitionAlloc] Don't clear the thread cache in MemoryReclaimer.
The thread cache should not be fully cleared by periodic reclaim in
MemoryReclaimer, as this is too aggressive, and there is already a
periodic purge limit within it.

Make sure that this only happens in low memory situations.

Bug: 998048, 1169159, 1169157
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2734622
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859818}

+++ base/allocator/partition_allocator/memory_reclaimer.cc
@@ -106,7 +106,7 @@ PartitionAllocMemoryReclaimer::~PartitionAllocMemoryReclaimer() = default;
 void PartitionAllocMemoryReclaimer::ReclaimAll() {
   constexpr int kFlags = PartitionPurgeDecommitEmptySlotSpans |
                          PartitionPurgeDiscardUnusedSystemPages |
-                         PartitionPurgeForceAllFreed;
+                         PartitionPurgeAggressiveReclaim;
   Reclaim(kFlags);
 }
 
@@ -131,14 +131,18 @@ void PartitionAllocMemoryReclaimer::Reclaim(int flags) {
   // the end of the remaining active slots.
   {
     using PCScan = internal::PCScan<internal::ThreadSafe>;
-    const auto invocation_mode = flags & PartitionPurgeForceAllFreed
+    const auto invocation_mode = flags & PartitionPurgeAggressiveReclaim
                                      ? PCScan::InvocationMode::kForcedBlocking
                                      : PCScan::InvocationMode::kBlocking;
     PCScan::Instance().PerformScanIfNeeded(invocation_mode);
   }
 
 #if defined(PA_THREAD_CACHE_SUPPORTED)
-  internal::ThreadCacheRegistry::Instance().PurgeAll();
+  // Don't completely empty the thread cache outside of low memory situations,
+  // as there is periodic purge which makes sure that it doesn't take too much
+  // space.
+  if (PartitionPurgeAggressiveReclaim)
+    internal::ThreadCacheRegistry::Instance().PurgeAll();
 #endif
 
   for (auto* partition : thread_safe_partitions_)
